Make gui_t 32-bit

+++ gui_t.hpp
@@ -3,7 +3,7 @@
 
 #include <cstring>
 
-using gui_t = size_t;
+using gui_t = uint32_t;
 
 
 #endif // __GUI_T__
